feat: move warning logic to util
* Clean up component by moving warning logic to utility function
* Renamed utility function
