+++ packages/app/src/components/page-information-block/page-information-block.tsx
@@ -1,6 +1,5 @@
 import css from '@styled-system/css';
 import { Warning } from '@corona-dashboard/icons';
-import { gmData, vrData } from '@corona-dashboard/common';
 import { isValidElement, ReactNode } from 'react';
 import styled from 'styled-components';
 import { ArticleSummary } from '~/components/article-teaser';
@@ -15,8 +14,7 @@ import { Header } from './components/header';
 import { Metadata, MetadataProps } from './components/metadata';
 import { PageLinks } from './components/page-links';
 import { WarningTile } from '~/components/warning-tile';
-import { getScopedGmWarning } from '~/utils/get-scoped-gm-warning';
-import { useIntl } from '~/intl';
+import { useScopedWarning } from '~/utils/use-scoped-warning';
 
 interface InformationBlockProps {
   title?: string;
@@ -53,21 +51,7 @@ export function PageInformationBlock({
   name,
   warning,
 }: InformationBlockProps) {
-  const { siteText } = useIntl();
-  const scopedGmName = siteText.gemeente_index.municipality_warning;
-  const scopedGm = gmData.find(
-    (gm) =>
-      gm.name === scopedGmName ||
-      (gm.searchTerms && gm.searchTerms.includes(scopedGmName))
-  );
-  const scopedVr = vrData.find((vr) => vr.code === scopedGm?.vrCode);
-
-  const scopedWarning = getScopedGmWarning(
-    scopedGmName,
-    name || '',
-    warning || '',
-    scopedVr?.name || ''
-  );
+  const scopedWarning = useScopedWarning(name || '', warning || '');
 
   const MetaDataBlock = metadata ? (
     <MetadataBox>
